// src/app/modules/operations/model/pur-inv-detail.ts
export class PurInvDetail {
    ID: number = 0;
    BuyIn: string = '';
    ItemID: number = 0;
    Cost: number = 0;
    Trade: number = 0;
    Sale: number = 0;
    Retail: number = 0;
    Ctn: number = 0;
    Pcs: number = 0;
    Pack: number = 0;
    DiscountPer: number = 0;
    DiscountAmt: number = 0;
    SalePer: number = 0;
    SaleAmt: number = 0;
    NetAmt: number = 0;
    ApplyDiscount: boolean = false;
    ApplySaleTax: boolean = false;
    Bonus: number = 0;
    TAmt: number = 0;
    // Hel
    Qty: number = 0;
  }
  